# Refactor Optimizer for correctness