--- a/gramps/gen/filters/__init__.py
+++ b/gramps/gen/filters/__init__.py
@@ -42,5 +42,10 @@ def reload_custom_filters():
     CustomFilters.load()
 
 
+def set_custom_filters(filter_list):
+    global CustomFilters
+    CustomFilters = filter_list
+
+
 # if not CustomFilters:  # moved to viewmanager
 # reload_custom_filters()