When the filter optimizer combines the result of two or more sub-filters, it is currently using the logical_op of the right hand sub-filter. Fix this so that the optimizer uses the parent filter logical_op

Fixes #13799 https://gramps-project.org/bugs/view.php?id=13799
Related forum discussion: https://gramps.discourse.group/t/strange-filter-behavior-in-6-0-1/7505/16

To reproduce using the example Gramps tree
1. create a new filter called "Ancestors of" with a single rule, "Ancestors of <person>"
set ID to I0044, Inclusive ticked
leave the Options set to "all rules must apply"
2. create a new filter called "Siblings of Ancestors" with a single rule, "Siblings of <filter> match"
set the filter to "Ancestors of"
leave the Options set to "all rules must apply"
3. Create a new filter called "The Family and their Spouces" with two rules
a. rule 1: People matching the <filter> and set the filter to "Ancestors of"
b. rule 2: People matching the <filter> and set the filter to "Siblings of Ancestors"
set the Options to "At least one rule must apply"

Filter the people view using the filter "Siblings of Ancestors". 25 people are expected. 0 are shown

The filter "Ancestors of" returns 7 people
The filter "Siblings of Ancestors" returns 18 people
So the union of these two results must, at a minimum, return 18 people. Therefore a result of 0 people cannot be correct.