@Nick-Hall 2025-04-25
We should include a unit test with a fix like this.