@dsblank 2025-04-26
It is probably because I altered the global CustomFilters. Let me see if I can cleanup after the new tests.