@dsblank 2025-04-26
Okay, I think I have made the pytest and Gramps gods happy.