@Nick-Hall 2025-04-26
Please add entries in the `po/POTFILES.skip` file for the new files.