@kkujansuu 2025-04-28
Sorry to intrude but as far as I can see filtering still does not work in Gramps 6 even with this optimizer fix. See for example the filters below.

The filter F1 obviously should yield all individuals in the database (since it is an "or" filter with the "Everyone" rule).
But it only yields the same results as filter F3:

```
  <object type="Person">
    <filter name="F1" function="or">
      <rule class="Everyone" use_regex="False" use_case="False">
      </rule>
      <rule class="MatchesFilter" use_regex="False" use_case="False">
        <arg value="F2"/>
      </rule>
    </filter>
    <filter name="F2" function="and">
      <rule class="MatchesFilter" use_regex="False" use_case="False">
        <arg value="F3"/>
      </rule>
    </filter>
    <filter name="F3" function="and">
      <rule class="IsAncestorOf" use_regex="False" use_case="False">
        <arg value="I0041"/>
        <arg value="0"/>
      </rule>
    </filter>
  </object>

```
![test](https://github.com/user-attachments/assets/1bc6189a-ae7f-4f95-aa5c-f9b8a28f58a9)

Or have I misunderstood something?