@dsblank 2025-04-28
Added @kkujansuu 's additional issue as another set of unittests. Now we have one failure to fix in the optimizer.