@stevenyoungs 2025-04-28
Here's a tabular version of the nice diagram @kkujansuu created. It includes additional information on the rules which have a `selected_handles` attribute.

|name|rule|logical_op|optimizer|
|--|--|--|--|
|F3|Ancestors of <person> "I0041"|and|selected_handles|
|F2|People matching the <filter> "F3"|and||
|F1|Everyone<br>People matching the <filter> "F2"|or||

The filter optimizer walks the tree of filters, looking for rules where
1. there is a `selected_handles` attribute
2. the logical_op is "and"

Where both conditions are true, the `selected_handles` are used to reduce the search space for any other, non-optimized rules. It reduces the search space by limiting handles which are checked in subsequent rules to those in `selected_handles` (`handles_in`) or for inverted rules, testing all handles in the db, excluding those handles in `selected_handles` (`handles_out`).

Earlier in this PR, I modified test 2 (`logical_op == 'and'`) to check the `logical_op` of the parent filter. The parent filter is F2 in this scenario. As can be seen, this also has a `logical_op` of `'and'`.

From the table above, we can see that the rule in filter F3 meets both conditions (`selected_handles` and `logical_op=='and'` [or `parent_logical_op=='and'`]). The optimizer therefore *only* considers the 16 handles resulting from filter F2 when evaluating the "Everyone" rule. Whilst this is quicker, it is also, unfortunately, wrong.

Alas, I'm still not sure what the correct fix is....