@dsblank 2025-04-29
@stevenyoungs, I made a change [cc8bc77](https://github.com/gramps-project/gramps/pull/2052/commits/cc8bc775777df1f37a4891ca9d54d68a14a9d968) that I believe fixes the issues, but not 100% sure yet. Need further testing.