@dsblank 2025-05-10
Although the optimization of `intersection` *only* reduced the time from milliseconds to nanoseconds, it was easy and could speedup could accumulate. I tested with large sets of handles, with an empty set at the end. Test code:

```python
import random
from functools import reduce
from gramps.gen.utils.id import create_id

def make_sets(n):
    sets = []
    for i in range(n):
        sets.append(set([create_id() for i in range(random.randint(20, 999_999))]))
    sets.append(set([]))
    return sets

sets = make_sets(10)
```

Without optimization:
```ipython
In [19]: %%time
    ...: functools.reduce(lambda x, y: x & y, sets, sets[0] if sets else set())
    ...: 
    ...: 
CPU times: user 46 ms, sys: 8.99 ms, total: 54.9 ms
Wall time: 54.9 ms
```

With optimization:
```ipython
In [20]: %%time
    ...: functools.reduce(lambda x, y: x & y, sorted(sets, key=lambda s: len(s)), sets[0] if sets else set())
    ...: 
    ...: 
CPU times: user 6 μs, sys: 1 μs, total: 7 μs
Wall time: 8.11 μs
```

The optimization of union could not make such large increases in speed, but I sped it quite a bit (4x) by using a trick to do more processing in C:

Without optimization:
```ipython
In [22]: %%time
    ...: x = reduce(lambda x, y: x | y, sets, set())
    ...: 
    ...: 
CPU times: user 1.79 s, sys: 653 ms, total: 2.45 s
Wall time: 2.45 s
```

With optimization:
```ipython
In [21]: %%time
    ...: x = set.union(*sets)
    ...: 
    ...: 
CPU times: user 289 ms, sys: 144 ms, total: 433 ms
Wall time: 433 ms
```

The optimization when you have an id_list was easy:

```python
        if id_list:
            optimizer = Optimizer(set(id_list), self)
        else:
            optimizer = Optimizer(set(self.get_all_handles(db)), self)
        possible_handles = optimizer.get_possible_handles()
```